+++ Demo/VFLRunner/Sources/VFLExampleViewController.swift
@@ -8,32 +8,14 @@
 import UIKit
 import VFL
 
-extension VFLExampleViewController {
+class VFLViewController: UIViewController {
+  private lazy var vfl = VFL(view)
+  private let contentView: UIView
   
-  class ContentView: UIView {
-    init(example: VFLExample, frame: CGRect = .zero) {
-      super.init(frame: frame)
-      let vws = (0..<example.viewCount).map { _ in UIView.create() }
-      let names = (0..<example.viewCount).map { "vw\($0)" }
-      VFL(self)
-        .add(subviews: vws, names: names)
-        .appendConstraints(formats: example.formats)
-    }
-    
-    required init?(coder: NSCoder) {
-      fatalError("init(coder:) has not been implemented")
-    }
-  }
-}
-
-class VFLExampleViewController: UIViewController {
-  let example: VFLExample
-  private var ctx = VFL()
-  
-  init(example: VFLExample) {
-    self.example = example
+  init(contentView: UIView, title: String) {
+    self.contentView = contentView
     super.init(nibName: nil, bundle: nil)
-    self.title = example.name
+    self.title = title
   }
   
   required init?(coder: NSCoder) {
@@ -42,20 +24,21 @@ class VFLExampleViewController: UIViewController {
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    let vw = ContentView(example: example)
-    ctx.setParent(view).add(subview: vw, name: "vw")
+    vfl
+      .add(subview: contentView, name: "contentVw")
+
     view.backgroundColor = .white
   }
   
   override func viewSafeAreaInsetsDidChange() {
     super.viewSafeAreaInsetsDidChange()
-    ctx.replaceConstraints(
-      metrics: view.safeAreaInsets.metrics,
-      formats: [
-        "V:|-(top)-[vw]-(bottom)-|",
-        "H:|-(left)-[vw]-(right)-|"
-      ]
-    )
+    vfl
+      .replaceConstraints(
+        metrics: view.safeAreaInsets.metrics,
+        formats: [
+          "V:|-(top)-[contentVw]-(bottom)-|",
+          "H:|-(left)-[contentVw]-(right)-|"
+        ]
+      )
   }
 }
-